requirements.txt
dj-database-url==0.5.0
dj-static==0.0.6
Django==2.2.9
django-filter==21.1
django-localflavor==3.1
django-pwa==1.0.10
django-smart-selects==1.5.9
django-stdimage==5.3.0
django-widget-tweaks==1.4.9
gunicorn==20.1.0
Pillow==8.4.0
psycopg2-binary==2.8.6
python-stdnum==1.17
pytz==2021.3
six==1.16.0
sqlparse==0.4.2
static3==0.7.0
